Rename the theme folders
